+++ src/ObjectCollection/BackedEnumCollection.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Steevanb\PhpCollection\ObjectCollection;
+
+/** @extends AbstractObjectCollection<\BackedEnum> */
+class BackedEnumCollection extends AbstractObjectCollection
+{
+    public static function getValueFqcn(): string
+    {
+        return \BackedEnum::class;
+    }
+}