+++ src/ObjectCollection/CountableNullableCollection.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Steevanb\PhpCollection\ObjectCollection;
+
+/** @extends AbstractObjectNullableCollection<\Countable|null> */
+class CountableNullableCollection extends AbstractObjectNullableCollection
+{
+    public static function getValueFqcn(): string
+    {
+        return \Countable::class;
+    }
+}