+++ src/ObjectCollection/DateTimeInterfaceNullableCollection.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Steevanb\PhpCollection\ObjectCollection;
+
+/** @extends AbstractObjectNullableCollection<\DateTimeInterface|null> */
+class DateTimeInterfaceNullableCollection extends AbstractObjectNullableCollection
+{
+    public static function getValueFqcn(): string
+    {
+        return \DateTimeInterface::class;
+    }
+}